Fix Error Reporting
